requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
dj-database-url==0.5.0
Django==3.0.5
django-crispy-forms==1.9.0
django-heroku==0.3.1
django-tinymce==3.0.2
django-tinymce4-lite==1.8.0
gunicorn==20.0.4
idna==2.9
jsmin==2.2.2
psycopg2==2.7.5
pytz==2019.3
requests==2.23.0
sqlparse==0.3.1
urllib3==1.25.9
whitenoise==5.0.1
